requirement.txt
torch
torchvision
torchaudio
numpy
pillow
matplotlib
gradio